lambda-code/monitoring-lambda-code/requirements.txt
boto3==1.34.69
requests==2.31.0
pyOpenSSL==24.0.0
httpx==0.27.0
